requirements.txt
azure-cognitiveservices-vision-computervision
keyboard
opencv-python
requests
python-dotenv
msrest